Remove get_pairings()
