+++ aiohomekit/__main__.py
@@ -147,7 +147,7 @@ async def discover(args):
 
 async def pair(args):
     async with get_controller(args) as controller:
-        if args.alias in controller.get_pairings():
+        if args.alias in controller.pairings:
             print(f'"{args.alias}" is a already known alias')
             return False
 
@@ -176,12 +176,12 @@ async def pair(args):
 async def get_accessories(args: Namespace) -> bool:
     async with get_controller(args) as controller:
 
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
         try:
-            pairing = controller.get_pairings()[args.alias]
+            pairing = controller.pairings[args.alias]
             data = await pairing.list_accessories_and_characteristics()
             controller.save_data(args.file)
         except Exception:
@@ -212,11 +212,11 @@ async def get_accessories(args: Namespace) -> bool:
 async def get_characteristics(args: Namespace) -> bool:
     async with get_controller(args) as controller:
 
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
-        pairing = controller.get_pairings()[args.alias]
+        pairing = controller.pairings[args.alias]
 
         # convert the command line parameters to the required form
         characteristics = [
@@ -250,12 +250,12 @@ async def get_characteristics(args: Namespace) -> bool:
 async def put_characteristics(args: Namespace) -> bool:
     async with get_controller(args) as controller:
 
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
         try:
-            pairing = controller.get_pairings()[args.alias]
+            pairing = controller.pairings[args.alias]
 
             # FIXME use Service.build_update
 
@@ -290,12 +290,12 @@ async def put_characteristics(args: Namespace) -> bool:
 
 async def identify(args: Namespace) -> bool:
     async with get_controller(args) as controller:
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
         try:
-            pairing = controller.get_pairings()[args.alias]
+            pairing = controller.pairings[args.alias]
             await pairing.identify()
         except Exception:
             logging.exception("Unhandled error whilst identifying device")
@@ -306,11 +306,11 @@ async def identify(args: Namespace) -> bool:
 
 async def list_pairings(args: Namespace) -> bool:
     async with get_controller(args) as controller:
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             exit(-1)
 
-        pairing = controller.get_pairings()[args.alias]
+        pairing = controller.pairings[args.alias]
         try:
             pairings = await pairing.list_pairings()
         except Exception as e:
@@ -332,11 +332,11 @@ async def list_pairings(args: Namespace) -> bool:
 
 async def remove_pairing(args):
     async with get_controller(args) as controller:
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
-        pairing = controller.get_pairings()[args.alias]
+        pairing = controller.pairings[args.alias]
         await pairing.remove_pairing(args.controllerPairingId)
         controller.save_data(args.file)
         print(f'Pairing for "{args.alias}" was removed.')
@@ -345,7 +345,7 @@ async def remove_pairing(args):
 
 async def unpair(args):
     async with get_controller(args) as controller:
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
@@ -357,11 +357,11 @@ async def unpair(args):
 
 async def get_events(args):
     async with get_controller(args) as controller:
-        if args.alias not in controller.get_pairings():
+        if args.alias not in controller.pairings:
             print(f'"{args.alias}" is no known alias')
             return False
 
-        pairing = controller.get_pairings()[args.alias]
+        pairing = controller.pairings[args.alias]
 
         # convert the command line parameters to the required form
         characteristics = [
