+++ aiohomekit/utils.py
@@ -1,5 +1,7 @@
 import enum
+import re
 
+from aiohomekit.exceptions import MalformedPinError
 from aiohomekit.model.characteristics import Characteristic
 
 
@@ -25,3 +27,15 @@ def clamp_enum_to_char(all_valid_values: enum.EnumMeta, char: Characteristic):
         valid_values = valid_values.intersection(set(char.valid_values))
 
     return valid_values
+
+
+def check_pin_format(pin: str) -> None:
+    """
+    Checks the format of the given pin: XXX-XX-XXX with X being a digit from 0 to 9
+
+    :raises MalformedPinError: if the validation fails
+    """
+    if not re.match(r"^\d\d\d-\d\d-\d\d\d$", pin):
+        raise MalformedPinError(
+            "The pin must be of the following XXX-XX-XXX where X is a digit between 0 and 9."
+        )
